move connect types to voiceConnect file
